Arduino NANO 33 BLE - init turn PWR LED on
Not a big deal, but a few of us have noticed that when we program the BLE boards
with the MBED version, the Power LED light is on.  When we we program it with
Zephyr it is off.

As the datasheets and Arduino pin pages such as Power and not USER, I believe
it should at least default to on.